@@ -6,11 +6,11 @@
 
     <groupId>org.camunda.community</groupId>
     <artifactId>zeebe-cherry-framework</artifactId>
-    <version>1.1.0</version>
+    <version>1.2.0</version>
 
     <properties>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
         <zeebe.version>8.0.3</zeebe.version>
         <junit.jupiter.version>5.8.2</junit.jupiter.version>
         <opensagres.version>2.0.3</opensagres.version>
@@ -49,6 +49,19 @@
             <version>2.9.0</version>
         </dependency>
 
+
+        <!-- Accept Camunda Connector -->
+        <dependency>
+            <groupId>io.camunda.connector</groupId>
+            <artifactId>connector-core</artifactId>
+            <version>0.1.0</version>
+        </dependency>
+        <dependency>
+            <groupId>io.camunda.connector</groupId>
+            <artifactId>connector-runtime-job-worker</artifactId>
+            <version>0.2.0-SNAPSHOT</version>
+        </dependency>
+
         <!-- JSON LocalDateTime -->
         <dependency>
             <groupId>com.fasterxml.jackson.datatype</groupId>
@@ -67,10 +80,12 @@
             <artifactId>spring-boot-starter-mustache</artifactId>
         </dependency>
 
-        <!-- Actuator -->
+
+        <!-- Process the configuration file -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-actuator</artifactId>
+            <artifactId>spring-boot-configuration-processor</artifactId>
+            <optional>true</optional>
         </dependency>
 
         <!-- Webjars -->
@@ -195,8 +210,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.10.1</version>
                 <configuration>
-                    <source>11</source>
-                    <target>11</target>
+                    <source>16</source>
+                    <target>16</target>
                 </configuration>
             </plugin>
         </plugins>
